requirements.txt
django==4.2.11
djangorestframework==3.15.1
psycopg2==2.9.9
celery[redis]==5.2.7
twilio==9.0.5